# Switching back to splice, to avoid bloat and garbage collection time