--- a/packages/bitcore-node/src/utils/partition.ts
+++ b/packages/bitcore-node/src/utils/partition.ts
@@ -1,4 +1,4 @@
 export function partition<T>(array: T[], n: number): T[][] {
   n = n > 0 ? Math.ceil(n) : 1;
-  return array.length ? [array.slice(0, n)].concat(partition(array.slice(n), n)) : [];
+  return array.length ? [array.splice(0, n)].concat(partition(array.splice(n), n)) : [];
 }