--- a/packages/bitcore-node/test/unit/utils/partition.unit.ts
+++ b/packages/bitcore-node/test/unit/utils/partition.unit.ts
@@ -13,28 +13,28 @@ describe('Partition', () => {
     let testArr = [1, 2, 3, 4, 5];
     let partitioned = partition(testArr, 0);
     expect(partitioned).to.deep.equal([[1], [2], [3], [4], [5]]);
-    expect(testArr).to.deep.equal([1, 2, 3, 4, 5]);
+    expect(testArr).to.deep.equal([]);
   });
 
   it('should handle one', () => {
     let testArr = [1, 2, 3, 4, 5];
     let partitioned = partition(testArr, 1);
     expect(partitioned).to.deep.equal([[1], [2], [3], [4], [5]]);
-    expect(testArr).to.deep.equal([1, 2, 3, 4, 5]);
+    expect(testArr).to.deep.equal([]);
   });
 
   it('should handle two', () => {
     let testArr = [1, 2, 3, 4, 5];
     let partitioned = partition(testArr, 2);
     expect(partitioned).to.deep.equal([[1, 2], [3, 4], [5]]);
-    expect(testArr).to.deep.equal([1, 2, 3, 4, 5]);
+    expect(testArr).to.deep.equal([]);
   });
 
   it('should handle between one and zero', () => {
     let testArr = [1, 2, 3, 4, 5];
     let partitioned = partition(testArr, 0.15);
     expect(partitioned).to.deep.equal([[1], [2], [3], [4], [5]]);
-    expect(testArr).to.deep.equal([1, 2, 3, 4, 5]);
+    expect(testArr).to.deep.equal([]);
   });
 
   it('should handle different sizes of arrays', () => {
@@ -50,7 +50,7 @@ describe('Partition', () => {
       if (!lastBatch) {
         console.error('Array partition fails with length', randomLen);
       }
-      expect(randomArr[randomArr.length - 1]).to.equal(lastBatch[lastBatch.length - 1]);
+      expect(randomArr).to.deep.equal([]);
     }
   });
 });