Memory usage is likely higher for 32 MB blocks due to not using splice. 
This MR reverts that, and makes the changes required to preserve the logic while using splice

# Before 
1028M/1569M+
```
Restting database
Generating blocks
Adding blocks
................................................................................
Benchmark for 80 (1 MB) blocks completed after 53.387 s
1.498492142281829 MB/s
0.6673375 Seconds/Block
Restting database
Generating blocks
Adding blocks
.....
Benchmark for 5 (32 MB) blocks completed after 88.812 s
1.801558347970995 MB/s
17.7624 Seconds/Block
```

# After 
```
Restting database
Generating blocks
Adding blocks
................................................................................
Benchmark for 80 (1 MB) blocks completed after 42.136 s
1.8986140117714068 MB/s
0.5267000000000001 Seconds/Block
Restting database
Generating blocks
Adding blocks
.....
Benchmark for 5 (32 MB) blocks completed after 88.958 s
1.7986015872659007 MB/s
17.7916 Seconds/Block
```